requirements.txt
# requirements.txt
# Основные зависимости для Excel Micro DB

# GUI Framework
PySide6>=6.7.0

# Работа с данными и Excel
numpy>=1.24.0
pandas>=2.0.3
openpyxl>=3.1.2

# Визуализация данных
matplotlib>=3.7.0

# Работа с конфигурацией YAML
PyYAML>=6.0

# Работа с переменными окружения
python-dotenv>=1.0.0